remove import pkcs11 from common.go
