+++ internal/peer/common/common.go
@@ -12,7 +12,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	"reflect"
 	"strings"
 	"time"
 
@@ -21,7 +20,6 @@ import (
 	pb "github.com/hyperledger/fabric-protos-go/peer"
 	"github.com/hyperledger/fabric/bccsp"
 	"github.com/hyperledger/fabric/bccsp/factory"
-	"github.com/hyperledger/fabric/bccsp/pkcs11"
 	"github.com/hyperledger/fabric/common/channelconfig"
 	"github.com/hyperledger/fabric/common/flogging"
 	"github.com/hyperledger/fabric/core/config"
@@ -165,7 +163,7 @@ func InitBCCSPConfig(bccspConfig *factory.FactoryOpts) error {
 	opts := viper.DecodeHook(mapstructure.ComposeDecodeHookFunc(
 		mapstructure.StringToTimeDurationHookFunc(),
 		mapstructure.StringToSliceHookFunc(","),
-		StringToKeyIds(),
+		factory.StringToKeyIds(),
 	))
 
 	if err := subv.Unmarshal(&bccspConfig, opts); err != nil {
@@ -206,38 +204,6 @@ func InitCrypto(mspMgrConfigDir, localMSPID, localMSPType string) error {
 	}
 
 	return nil
-}
-
-// StringToKeyIds returns a DecodeHookFunc that converts
-// strings to pkcs11.KeyIDMapping.
-func StringToKeyIds() mapstructure.DecodeHookFunc {
-	return func(
-		f reflect.Type,
-		t reflect.Type,
-		data interface{}) (interface{}, error) {
-		if f.Kind() != reflect.String {
-			return data, nil
-		}
-
-		if t != reflect.TypeOf(pkcs11.KeyIDMapping{}) {
-			return data, nil
-		}
-
-		res := pkcs11.KeyIDMapping{}
-		raw := data.(string)
-		if raw == "" {
-			return res, nil
-		}
-
-		rec := strings.Fields(raw)
-		if len(rec) != 2 {
-			return res, nil
-		}
-		res.SKI = rec[0]
-		res.ID = rec[1]
-
-		return res, nil
-	}
 }
 
 // SetBCCSPKeystorePath sets the file keystore path for the SW BCCSP provider
